requirements.txt
mcp>=0.1.6
mcp-cli>=0.1.4
uvicorn
psycopg2-binary
